+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,11 @@
+package duke;
+
+/**
+ * A checked exception class, personalized to fit the Duke application.
+ */
+public class DukeException extends Exception {
+    public DukeException(String a) {
+        super(a);
+    }
+}
+
