--- src/main/java/duke/Gui.java
@@ -1,134 +1,0 @@
-package duke;
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-
-/**
- * Gui handles setting the stage for the GUI. and responding to user input.
- */
-public class Gui extends Application {
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
-    private Parser parser;
-    private Storage storage;
-    private TaskList taskList;
-
-
-    /**
-     * Default constructor for Gui class, initializes the parser and Tasklist.
-     */
-    public Gui() {
-        DtFormat dtf = new DtFormat();
-        Ui ui = new Ui(dtf);
-        storage = new Storage("./data/");
-
-        try {
-            taskList = new TaskList(storage.load(), dtf);
-        } catch (DukeException e) {
-            ui.showLoadingError();
-            taskList = new TaskList();
-        }
-        CommandHandler ch = new CommandHandler(taskList, ui, dtf);
-        this.parser = new Parser(dtf, ui, taskList, ch);
-    }
-
-    /**
-     * Sets the stage for the GUI.
-     * @param stage the stage upon which to set the scene.
-     */
-
-    @Override
-    public void start(Stage stage) {
-
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
-
-        userInput = new TextField();
-        sendButton = new Button("Send");
-
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-        stage.show();
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-
-        mainLayout.setPrefSize(400.0, 600.0);
-
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(325.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput, 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
-
-    }
-
-    /**
-     * Function to handle user text input.
-     */
-    private void handleUserInput() {
-        String userText = userInput.getText();
-        String dukeText = getResponse(userInput.getText());
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(userText, user),
-                DialogBox.getDukeDialog(dukeText, duke));
-        userInput.clear();
-    }
-
-    /**
-     * Returns a response given a user string input.
-     */
-    public String getResponse(String input) {
-        DukeException ex = null;
-        try {
-            String ret = parser.parse(input);
-            storage.save(taskList.getItems());
-            return ret;
-        } catch (DukeException e) {
-            ex = e;
-        }
-        return ex.toString();
-    }
-
-}
