+++ src/main/java/duke/Launcher.java
@@ -1,0 +1,19 @@
+package duke;
+import javafx.application.Application;
+
+/**
+ * Entry point of the Duke program.
+ */
+
+public class Launcher {
+
+    /**
+     * Runs the Duke program, beginning the loop of receiving user
+     * input until the user terminates.
+     */
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
+
+
